% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_iso_amu.R
\name{get_iso_amu}
\alias{get_iso_amu}
\title{get_iso_amu.}
\usage{
get_iso_amu(
  x,
  isotopes = data.frame(isotope = c("198Hg", "32S"), mass = c(197.999, 31.995))
)
}
\arguments{
\item{x}{character.}

\item{isotopes}{Two column dataframe with isotope definitions.}
}
\value{
A single numeric value (0 in case that no isotope could be identified).
}
\description{
\code{get_iso_amu} will take a string and try to identify an
isotope name contained within. It will return the amu for this isotope.
}
\examples{
IsoCor:::get_iso_amu(x="198Hg")
IsoCor:::get_iso_amu(x="198Hg_corr")
IsoCor:::get_iso_amu(x="X_32S_corr")
IsoCor:::get_iso_amu(x="15S")

}
\keyword{internal}
