% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_the_user.R
\name{help_the_user}
\alias{help_the_user}
\title{help_the_user}
\usage{
help_the_user(filename)
}
\arguments{
\item{filename}{Name of the file as string (if necessary, containing also path).}
}
\value{
Returns the help text as HTML (opens a modal with helpt text as side effect).
}
\description{
Help Window: opens a modal with respective Help text (rendered from Rmd source file) for users.
}
\keyword{internal}
