% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_ratio.R
\name{iso_ratio}
\alias{iso_ratio}
\title{iso_ratio}
\usage{
iso_ratio(
  data = data.frame(X = rnorm(10), Y = rnorm(10)),
  method = c("mean", "area", "slope")[1],
  thr = 1
)
}
\arguments{
\item{data}{data.frame with two columns specifying data for isotope 1 and 2 respectively.}

\item{method}{Method to calculate the isotope ratio.}

\item{thr}{Threshold between 0..1 to limit the peaks scans used in the calculation (1=all scans, 0=apex only).}
}
\value{
A single numeric value given the robsut ratio estimate calculated from data.
}
\description{
\code{iso_ratio} will calculate a robust estimate of an
isotopic ratio between intensity values of 2 mass traces.
}
\details{
tbd.
}
\keyword{internal}
