% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clipboard.R
\name{read_clipboard_UI}
\alias{read_clipboard_UI}
\title{read_clipboard}
\usage{
read_clipboard_UI(id)
}
\arguments{
\item{btn_txt}{btn_txt.}

\item{n_rows}{\enumerate{
\item 
}}

\item{n_cols}{\enumerate{
\item 
}}
}
\value{
A data.frame.
}
\description{
\code{read_clipboard} will .
}
\details{
tbd
}
\examples{
\dontrun{
shinyApp(
 ui = shiny::fluidPage(read_clipboard_UI(id = "test")), 
 server = function(input, output, session) { 
   tmp <- read_clipboard_Server(id = "test") 
   observeEvent(tmp$d, print(tmp$d))
 }
)
}

}
\keyword{internal}
