% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\docType{data}
\name{testdata}
\alias{testdata}
\title{testdata.}
\format{
A list of 3 exp spectra files converted by read_raw_data.
}
\source{
dariya.tukhmetova@bam.de
}
\usage{
data(testdata)
}
\description{
testdata.
}
\keyword{datasets}
