% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_data.R
\docType{data}
\name{testdata_IDMS}
\alias{testdata_IDMS}
\title{testdata_IDMS.}
\format{
A list of 3 txt spectra files converted by read_raw_data.
}
\source{
dariya.tukhmetova@bam.de
}
\usage{
data(testdata_IDMS)
}
\description{
testdata_IDMS.
}
\keyword{datasets}
