% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_peak_boundaries.R
\name{find_peak_boundaries}
\alias{find_peak_boundaries}
\title{find_peak_boundaries}
\usage{
find_peak_boundaries(int = NULL, p = which.max(int), k = 3, min_scans = 3)
}
\arguments{
\item{int}{Numeric vector (of intensity valus).}

\item{p}{Index of peak position (usually 'which.max(int)).}

\item{k}{Number of scans at peak boarder to confirm peak valley.}

\item{min_scans}{Minimum number of scans in front or tail.}
}
\value{
A numeric vector of length 2 giving the indexes of peak start and peak end.
}
\description{
\code{find_peak_boundaries} will find peak boundaries, based on curve derivative.
}
\details{
tbd.
}
\examples{
x <- sin(seq(-pi,2*pi,0.01))+1
plot(x)
abline(v=IsoCor:::find_peak_boundaries(x))

}
\keyword{internal}
