% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_app.R
\name{ic_app}
\alias{ic_app}
\title{ic_app}
\usage{
ic_app()
}
\value{
A shiny app object.
}
\description{
\code{ic_app} will start a shiny app that allows to upload raw
data, process selectively and analyze different methods of ratio calculation
between two intensity traces.
}
\details{
The app is described in detail in
<\href{https://jali.shinyapps.io/IsoCor}{this publication}>.
}
\references{
\url{https://jali.shinyapps.io/IsoCor}
}
\seealso{
\link{iso_ratio}
}
