% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotopes.R
\docType{data}
\name{isotopes}
\alias{isotopes}
\title{isotopes.}
\format{
A data.frame of chemical isotopes.
}
\source{
R-package enviPat
}
\usage{
data(isotopes)
}
\description{
isotopes.
}
\keyword{datasets}
