% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw_data.R
\name{read_raw_data}
\alias{read_raw_data}
\title{read_raw_data}
\usage{
read_raw_data(path, format = "exp")
}
\arguments{
\item{path}{Valid filepath.}

\item{format}{Character, currently only 'exp' files are supported.}
}
\value{
A data.frame.
}
\description{
\code{read_raw_data} will import data in exp file format.
}
\details{
tbd.
}
\keyword{internal}
